third_party/WebKit/LayoutTests/accessibility/draw-focus-if-needed-expected.txt

PASS axButton1.x is 50
PASS axButton1.y is 50
PASS axButton1.width is 200
PASS axButton1.height is 100
PASS axContainer.x is 50
PASS axContainer.y is 200
PASS axContainer.width is 200
PASS axContainer.height is 100
PASS axButton2.x is 50
PASS axButton2.y is 200
PASS axButton2.width is 200
PASS axButton2.height is 100
PASS successfullyParsed is true

TEST COMPLETE
